Gate the reviewed D31 resource configuration boundary
Add versioned resource framing, preserve uint32 admission identity, and reject unsupported profiles and zero input bounds at configuration installation. Keep singleton admission separate.

Reviewed with Claude Code; follow-up tests cover identity mismatch, encoded specials and the real valid_config_data call. Include manually rebuilt negative-control evidence and source hashes.

D31 remains incomplete: this commit closes only the configuration boundary. Whole-input admission, complete limit compatibility, and live multi-account execution are not delivered here.

+++ crypto/block/workchain-input-admission.h
@@ -34,7 +34,7 @@ struct InputPolicyIdentity {
   std::int64_t engine_selector;
   std::uint64_t vm_mode;
   std::uint32_t descriptor_version;
-  std::uint16_t admission_version;
+  std::uint32_t admission_version;
 };
 
 class CandidateAdmissionSession;
